requirements.txt
langchain==0.3.21
langchain-core==0.3.20
langchain-community==0.3.20
langchain-text-splitters==0.3.7
transformers
torch
sentence-transformers
rank-bm25
beautifulsoup4
biopython
faiss-cpu
nltk
pymed
spacy
scikit-learn
